End of Sprint 4 (#10)
* Keeping up to date with python requirements.

* Update python framework necassary to populate intersection database.
The readme outlines the major python scripts that need to be ran in order to populate the intersection database 'alpha'.
Major changes include changing the num_oneway from a float that is a ratio of one way lanes, to the actual number of one way lanes.
Also the network used for all graph operations now are drive_service, which includes drive networks as well as service roads like entrances to businesses.
This code relies on the OSMNX framework, as well as MAPBOX. In order to use the MAPBOX API, one needs a token and the free limit is something like 50k/month.

* Added the intersection create query to the repo.
This is the query used to create the intersection table 'beta', and should be expanded on to include other info as well. It should also be added to the osm2pgl (or a newly created 'database-preparation'  service dockerfile once it is in a decent shape.

* Added requirements.txt to run folder (from old)

* Updated intersection creation query, deduplicated, added auto-counting id

* Added windows notes to README

* Added SQL file demoing duplicate OSM ID issue

* Fixed issue with temp file permissions.

* PEP8 cleanup on prepare_intersection_masks.py

* Moved python files from run to services/intersections. Added intersections_init file to run them. Minor pep-8 cleanup.

* Moved README and env_blank from run to services/intersections.

* Added ug-intersectionsgraph to devsecops; moved requirements.txt to that folder.

* Updated Dockerfile

* Turned temp file deletion back on.

* Updates from Dylan for PR of ticket DRWS-52
Changes include:
adding lines in the init.sh to incorporate changes in ticket DRWS-52.
changing .env_blank to include dummy values that are already in the repo also adding HOSTNAME variable.
Added some explicit instructions to the README on creating the .env
Remove unnecassary warnings in calculate_total_num_legs.py
Added a .gitkeep to the data directory that must exist in order for the masks/images to save.
Added a comment to intersections_init.py
Changed the path that is saved in the load_intersections.py file to reflect the location of the masks/images on the local machine. It doesn't matter now, as it will eventually be an S3 path, but this makes it easier to demo to customer.

* added postgres launch script and dockerfile

* Populated stub dockerfile to facilitate running postgres script

* Added hostname to psql call.

* Moved query under create-intersections-xsect folder

* Added service to docker compose.

* Added technical process document.

* Added dependency on osm2postgres's successful completion to intersectionsgraph-xsect

* Renamed intersectionsgraph-xsect to intersections-xsect.

* Changed permissions

* removed hardcoded values in favor of input from dockerfile for intersection-graph init

* Add crash model prototype.
This can also be used to export data for casse in the form of integrated csv files.

* Notebook used to produce the GLM model.

* Adding line to index crash data on spatial index.

* Reduce workers to 4.
For systems with large number of computer nodes, the postgres database will deny access to more than 10 threads or so. 4 threads are plenty.

* Tidbit on vresions.

* Process data on init (#7)

* Updated intersection creation query to be more explicit in its field types and primary key.

* Added table creation scripts to ease use of nodes, ways, and featuresd.

* Added dockerfile and moved SQL to osm directory under 'postprocess' subdirectory.

* Updated dockerfile and addd it to the composefile.

* Added postprocess pilot script.

* Added 'if not exists' checks to table creation.

* Added defensive checks to avoid double insertion.

* Recommented drop table in way_feature creation script.

* Updated docker compose to use correct image. Updated dockerfile to sleep less.

* Added cascade to way feature creation's commented drop table, so it drops instead of failing.

* Changing execute permissions on file.

* Renamed intersection creation query temporarily.

* And renamed it back.

* Making the graph service wait for the osmpostprocessing to finish.
This makes it a little easier on the memory requirements.

Co-authored-by: Mark Villarrubia <mark.villarrubia@intellibridge.us>
Co-authored-by: Dylan.Frizzell <dylan.frizzell@DylanFrizzellugroup-C02YWE17LVDC.local>

Co-authored-by: Dylan.Frizzell <dylan.frizzell@DylanFrizzellugroup-C02YWE17LVDC.local>
Co-authored-by: Mark Villarrubia <mark.villarrubia@intellibridge.us>
Co-authored-by: markv-ugroup <93332033+markv-ugroup@users.noreply.github.com>
Co-authored-by: goldenBoard <dylan.frizzell@ou.edu>